Remove WASM Serializer stuff
